customer や user や customer-setting 関連のモジュールを全て見直し
cf. #97
