+++ components/manage/content/news.vue
@@ -14,7 +14,7 @@ const emit = defineEmits<{
   remove: [id: number]
 }>()
 
-const { customerId } = useFoundation()
+const { customerId } = useCustomer()
 const apiKind = getNewsKind()
 
 const modal = ref(false)
