+++ components/publish/home/type1/contact.vue
@@ -1,5 +1,6 @@
 <script setup lang="ts">
-const { customerId, canEdit } = useFoundation()
+const { customerId } = useCustomer()
+const { canEdit } = useCustomerPageContext()
 
 const {
   contactRef,
