+++ components/publish/home/type1/information.vue
@@ -1,5 +1,6 @@
 <script setup lang="ts">
-const { customerId, canEdit } = useFoundation()
+const { customerId } = useCustomer()
+const { canEdit } = useCustomerPageContext()
 const {
   informationRef,
   onLoad,
