+++ components/publish/home/type1/newses.vue
@@ -2,7 +2,8 @@
 import type { NewsType } from '@/types/content'
 import { formatLocalDate } from '@/utils/misc'
 
-const { customerId, canEdit } = useFoundation()
+const { customerId } = useCustomer()
+const { canEdit } = useCustomerPageContext()
 const {
   filter,
   sort,
