+++ types/API/customer-setting-api.ts
@@ -1,4 +1,28 @@
-import type { SectionKind } from '@/types/customer-setting'
+import type {
+  ColorTheme,
+  DesignTheme,
+  SectionKind,
+  NetworkServiceink,
+} from '@/types/customer-setting'
+
+export type CustomerSettingApi = {
+  id: number
+  customerId: number
+  availContentsKind: string[]
+  pageTitle?: {
+    title: string
+    iconUrl?: string | null
+  }
+  colorTheme: ColorTheme
+  designTheme: DesignTheme
+  homeLayout?: {
+    kind: SectionKind
+    title: string
+    menuTitle?: string | null
+  }[]
+  snsLinks?: NetworkServiceink[] | null
+  accessSource?: string | null
+}
 
 export type PageSectionApi = {
   baseId: string
