+++ types/auth.d.ts
@@ -26,14 +26,6 @@ export type LoginForm = {
 }
 
 /**
- * Change Password Form
- */
-export type ChangePasswordFrom = {
-  password: string
-  passwordConfirm: string
-}
-
-/**
  * Login User 情報
  */
 export type LoginUser = CustomerUser
